alpaca_paper/llama/requirements.txt
flask
redis
Flask-SQLAlchemy
Flask-Migrate
sqlalchemy-json
flask-cors
psycopg2
sqlalchemy-utils
python-dotenv
requests